LZMR-47 implemented service interface
